Added Golang binary tree.
This is kinda cool I think as it lets you just have a tiny,
straightforward binary tree implementation with insertion, deletion, and
searching in a totally not thread safe environment. If you're going to
use this tree I suggest keeping the size down and not doing anything
mental like sharing it between several processes.

Need to implement the interface functions to use this for your data
type:
LessThan(other <of the same type>) bool
GreaterThan(other <of the same type>) bool
EqualTo(other <of the same type>) bool