Dat Class Structure.
Does Java even like recursive types? I know recursion really isn't a
popular thing in most of imperative land but when compared to faffing
around with a static array or linked listed of nodes this seems an awful
lot simpler.

Is it just me?

Also still getting used to how Java is structured..