Comments Added and Slight Refactor.
Comments ! WE HAVE CODE COMMENTS! .. About friggin time too.

Refactored some of the code to be a bit clean up and approach a more
Java-zen state. i.e. There are now getters and setters. *ahem*

I'm not convinced the Delete operation is correct. The handling of the
deletion of a node that has children that has children I don't think is
properly accounted for. I will need to test for this!