Refactored Java BTree, Added Java HashTable.
Refactored the Java Binary tree for a bit more Java-ness.

Implemented a super basic Java HashTable that has linear probing and a
rudimentary hashing function that sums the integer value of the key
characters and applies a modulo of the length of the storage table.

Currently there is no support for extending and then rebuilding the table.